% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOtable.R
\docType{data}
\name{ISOtable}
\alias{ISOtable}
\title{A lookup table for ISO codes and names}
\format{A tibble with 3 columns and 288 rows containing
ISO labels, ISO codes, and a label indicating the
level of aggregation.
\describe{
  \item{ISOlabel}{ISO area name (character string).}
  \item{ISO}{Three digit ISO code (integer)}
  \item{LocTypeName}{What level of aggregation the code pertains to (character strng)}
}}
\source{
\url{https://population.un.org/wpp/Download/Metadata/Documentation/}
}
\usage{
ISOtable
}
\description{
This table might be helpful for matching ISO
codes to location names (all levels of geography).
}
\keyword{datasets}
